move two utils from hmfs/hmf to utils/utilities to avoid circular imports in scripts
