+++ lmfdb/utils/__init__.py
@@ -11,6 +11,7 @@ __all__ = ['request', 'make_response', 'flash', 'url_for', 'render_template',
            'format_percentage', 'signtocolour', 'rgbtohex', 'pol_to_html',
            'web_latex', 'web_latex_ideal_fact', 'web_latex_split_on',
            'web_latex_split_on_pm', 'web_latex_split_on_re', 'display_knowl',
+           'teXify_pol', 'add_space_if_positive',
            'bigint_knowl', 'too_big', 'make_bigint', 'bigpoly_knowl',
            'factor_base_factor', 'factor_base_factorization_latex',
            'polyquo_knowl', 'web_latex_poly', 'list_to_latex_matrix',
@@ -49,6 +50,7 @@ from .utilities import (
     splitcoeff, comma, format_percentage, signtocolour, rgbtohex, pol_to_html,
     web_latex, web_latex_ideal_fact, web_latex_split_on, web_latex_split_on_pm,
     web_latex_split_on_re, display_knowl, bigint_knowl, too_big, make_bigint,
+    teXify_pol, add_space_if_positive,
     bigpoly_knowl, factor_base_factor, factor_base_factorization_latex,
     polyquo_knowl, web_latex_poly, list_to_latex_matrix, code_snippet_knowl,
     Pagination,
